requirements.txt
python-telegram-bot
google-api-python-client
googletrans==3.1.0a0
pytube
pytz
TikTokApi
instaloader
yt-dlp